run telemetry from dashboard
